Break up capi/__init__ into __init__, core, and error
