+++ openmc/capi/material.py
@@ -5,7 +5,8 @@ from weakref import WeakValueDictionary
 import numpy as np
 from numpy.ctypeslib import as_array
 
-from openmc.capi import _dll, _error_handler, NuclideView
+from . import _dll, NuclideView
+from .error import _error_handler
 
 
 __all__ = ['MaterialView', 'materials']
@@ -58,6 +59,7 @@ class MaterialView(object):
 
     """
     __instances = WeakValueDictionary()
+
     def __new__(cls, *args):
         if args not in cls.__instances:
             instance = super().__new__(cls)
